Add files for proper API server with hashing
The new headless API server includes hashing and salting and has
been tested - took some debugging to realize that the passwords
need to be converted to string before being passed to the
crypto.pbkdf2Sync() function. Now all is set and there is a very
good working system which is working with mongodb and saving things
in it.