Read labels from external file, added tests
The flask app is more scalable and modular by taking everything
from an external set of files - labels.txt has the model labels
and output_graph.pb has the tensorflow model object. This simple
abstraction will ensure that the system can be scaled to any
model which is trained on tensorflow.